Created snake.js file
Moved the snake speed var from game..js
and put it in snake.js and set it to export.
- updated game.js to import the snake speed...

Created snake.js file to denote all things for the snake
Created 2 funtions for updating the snake, and to draw the snake.
Set them to export and updated the import
on game.js to include them both.

Created funtions on game.js to call the
funtions from snake.js both functions are
basically empty just printing text to the
console at the moment.