Created Food
created food.js file
created a var food and set co-ordinates to
position it.
copied most code from snake .js
added import update draw food to game.js
Created Var expansion rate to defive a growth value for each time the snake eats
food.

created a expand snake function that is
using the expansion rate to,,, you
guessed it, expand the snake.
Created a function called onSnake to
identify if the snake is on the food aka
eating it.
The expansion is called for from the export
function update along with and update for
the food co-ordinates that are set statically
at the moment.

food.js
import onsnake and expand snake