[BUG] snake wont stop growing
console.log(newSegments) to help debug
Found at start of game a continuous log of '0'
occurs, when move snake it still logs '0'.
When snake hits food it starts logging '1'
it needs to revert back to '0' after using '1'
once so it only grows once.
[FIX] inside the addSegments func
this now resets the newSegments value to '0'
at the end of the function.
Snake now grows from 1 to 2 segments...
Commented out the console log for now...