Randomise food position
Created a func getRandomFoodPosition
that will loop while newFoodPosition is null or
onsnake is on newFoodPosition and returns a
newFoodPosition that is equal to a random
value defined by a function called
randomGridPosition [yet to be created]

Changed the food export update func
From static food position to the
getRandomFoodPosition function.

For now I am leaving the food start position static.

Wiil commit this and create the function
randomGridPosition next...