Create a game over
game.js
created func called checkDeath
which sets a var gameOver i the head of the
snake is outside of the grid or if the snake
intersects itself.

Set a var for gameOver and set initialise as false

Inside the main function now starts with if
gameOver return a message you loose.

grid.js
created function outsideGrid

snake.js
created funtion getSnakeHead
to get the array position 0 of snakeBody.
Created function snakeIntersection to check
if the snake has intersected itself.

game.js
import update with getSnakeHead - snakeIntersection from snake.js
import update with outsideGrid from grid.js

This has a bug now and wont gameover when
out of grid, only on snakeIntersection. I am
proceeding with this commit to avoid breaking
it more than I have : ) to be continued...